+++ src/plugins/intel_gpu/src/graph/impls/common/condition.cpp
@@ -4,7 +4,7 @@
 
 #include "condition_inst.h"
 #include "data_inst.h"
-#include "implementation_map.hpp"
+#include "impls/registry/implementation_map.hpp"
 #include "register.hpp"
 
 #include <algorithm>
