+++ src/plugins/intel_gpu/src/graph/impls/cpu/non_max_suppression.cpp
@@ -6,7 +6,7 @@
 #include "primitive_inst.h"
 #include "register.hpp"
 #include "cpu_impl_helpers.hpp"
-#include "implementation_map.hpp"
+#include "impls/registry/implementation_map.hpp"
 
 #include <vector>
 #include <queue>
