+++ src/plugins/intel_gpu/src/graph/impls/cpu/read_value.cpp
@@ -3,9 +3,8 @@
 //
 
 #include "read_value_inst.h"
-#include "implementation_map.hpp"
+#include "impls/registry/implementation_map.hpp"
 #include "register.hpp"
-#include "intel_gpu/runtime/error_handler.hpp"
 
 namespace cldnn {
 namespace cpu {
