+++ src/plugins/intel_gpu/src/graph/impls/cpu/scatter_update.cpp
@@ -4,9 +4,7 @@
 
 #include "register.hpp"
 #include "scatter_update_inst.h"
-#include "implementation_map.hpp"
-
-#include "intel_gpu/runtime/error_handler.hpp"
+#include "impls/registry/implementation_map.hpp"
 
 #include "openvino/op/scatter_update.hpp"
