+++ src/plugins/intel_gpu/src/graph/impls/onednn/deconvolution_onednn.cpp
@@ -7,7 +7,7 @@
 #include "impls/onednn/utils.hpp"
 #include "quantize_inst.h"
 #include "primitive_onednn_base.h"
-#include "implementation_map.hpp"
+#include "impls/registry/implementation_map.hpp"
 
 #include "impls/ocl/kernel_selector_helper.h"
