+++ src/plugins/intel_gpu/src/graph/impls/onednn/gemm_onednn.cpp
@@ -4,7 +4,7 @@
 
 #include "gemm_inst.h"
 #include "primitive_onednn_base.h"
-#include "implementation_map.hpp"
+#include "impls/registry/implementation_map.hpp"
 
 #include "kernel_selector_common.h"
