+++ src/plugins/intel_gpu/src/graph/impls/onednn/reorder_onednn.cpp
@@ -5,7 +5,7 @@
 #include "impls/onednn/utils.hpp"
 #include "reorder_inst.h"
 #include "primitive_onednn_base.h"
-#include "implementation_map.hpp"
+#include "impls/registry/implementation_map.hpp"
 
 #include "kernel_selector_common.h"
