+++ src/plugins/intel_gpu/src/graph/input_layout.cpp
@@ -4,7 +4,6 @@
 #include "input_layout_inst.h"
 #include "primitive_type_base.h"
 #include "intel_gpu/runtime/memory.hpp"
-#include "intel_gpu/runtime/error_handler.hpp"
 #include "json_object.h"
 #include <string>
 #include <memory>
