+++ src/plugins/intel_gpu/src/graph/scaled_dot_product_attention.cpp
@@ -5,7 +5,6 @@
 #include "scaled_dot_product_attention_inst.h"
 
 #include "primitive_type_base.h"
-#include "intel_gpu/runtime/error_handler.hpp"
 #include "json_object.h"
 #include <string>
 #include <vector>
